Extract `printIntersectionType`  `printUnionType`
